♻️ refactor: refactor user management logic by moving adminThunk to a dedicated user directory
